separate models from process files
